Rename logging module to log_commons to avoid namespace conflicts
