+++ analysis/deep2_r23_o32/zoom_and_gauss_general.py
@@ -22,7 +22,7 @@ from Metallicity_Stack_Commons.analysis.fitting import movingaverage_box1D, rms_
 from Metallicity_Stack_Commons import lambda0, line_name, line_type
 from Metallicity_Stack_Commons.column_names import filename_dict
 
-from .logging import log_stdout
+from .log_commons import log_stdout
 
 con1 = 3728.91 / 3726.16
